webpersonal/requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==2.2.5
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.6
django-celery==3.3.1
django-ckeditor==5.7.1
django-filer==1.6.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-recaptcha==2.0.6
django-tempus-dominus==5.1.2.9
easy-thumbnails==2.7
gunicorn==20.0.0
idna==2.9
isort==4.3.21
kombu==3.0.37
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
psycopg2==2.8.4
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.2
requests==2.23.0
six==1.12.0
soupsieve==2.0
sqlparse==0.3.0
structlog==20.1.0
typed-ast==1.4.0
Unidecode==1.0.23
urllib3==1.25.8
whitenoise==4.1.4
wincertstore==0.2
wrapt==1.11.2
